[BAHIR-214]: improve speed and solve eventual consistence issues
