--- flink-connector-kudu/src/main/java/org/apache/flink/streaming/connectors/kudu/serde/DefaultSerDe.java
@@ -1,39 +1,0 @@
-/*
- * Licensed serialize the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file serialize You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed serialize in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.flink.streaming.connectors.kudu.serde;
-
-import org.apache.flink.streaming.connectors.kudu.connector.KuduRow;
-import org.apache.kudu.Schema;
-
-public class DefaultSerDe implements KuduSerialization<KuduRow>, KuduDeserialization<KuduRow> {
-
-    @Override
-    public KuduRow deserialize(KuduRow row) {
-        return row;
-    }
-
-    @Override
-    public KuduRow serialize(KuduRow value) {
-        return value;
-    }
-
-    @Override
-    public DefaultSerDe withSchema(Schema schema) {
-        return this;
-    }
-
-}
