--- flink-connector-kudu/src/test/java/org/apache/flink/streaming/connectors/kudu/serde/PojoSerDeTest.java
@@ -1,57 +1,0 @@
-/*
- * Licensed serialize the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file serialize You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed serialize in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.flink.streaming.connectors.kudu.serde;
-
-import org.apache.flink.streaming.connectors.kudu.connector.KuduColumnInfo;
-import org.apache.flink.streaming.connectors.kudu.connector.KuduRow;
-import org.apache.flink.streaming.connectors.kudu.connector.KuduTableInfo;
-import org.apache.kudu.Type;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class PojoSerDeTest {
-
-    public class TestPojo {
-        private String field1;
-        private String field2;
-        private String field3;
-
-        public TestPojo() {
-            field1 = "field1";
-            field2 = "field2";
-            field3 = "field3";
-        }
-    }
-
-    @Test
-    public void testFieldsNotInSchema() {
-
-        TestPojo pojo = new TestPojo();
-
-        KuduTableInfo tableInfo = KuduTableInfo.Builder.create("test")
-                .addColumn(KuduColumnInfo.Builder.create("field1", Type.STRING).key(true).hashKey(true).build())
-                .addColumn(KuduColumnInfo.Builder.create("field2", Type.STRING).build())
-                .build();
-
-        KuduRow row = new PojoSerDe<>(TestPojo.class).withSchema(tableInfo.getSchema()).serialize(pojo);
-
-        Assertions.assertEquals(2, row.blindMap().size());
-        Assertions.assertEquals("field1", row.getField("field1"));
-        Assertions.assertEquals("field2", row.getField("field2"));
-
-    }
-}
